+++ qa/rpc-tests/mergetoaddress_sapling.py
@@ -14,7 +14,10 @@ class MergeToAddressSapling (BitcoinTestFramework):
         self.helper.setup_chain(self)
 
     def setup_network(self, split=False):
-        self.helper.setup_network(self, ['-anchorconfirmations=1'])
+        self.helper.setup_network(self, [
+            '-anchorconfirmations=1',
+            '-allowdeprecated=legacy_privacy',
+        ])
 
     def run_test(self):
         self.helper.run_test(self)