+++ src/wallet/asyncrpcoperation_mergetoaddress.h
@@ -9,8 +9,10 @@
 #include "asyncrpcoperation.h"
 #include "primitives/transaction.h"
 #include "transaction_builder.h"
+#include "uint256.h"
 #include "wallet.h"
 #include "wallet/paymentdisclosure.h"
+#include "wallet/wallet_tx_builder.h"
 #include "zcash/Address.hpp"
 #include "zcash/JoinSplit.hpp"
 
@@ -25,16 +27,8 @@
 
 using namespace libzcash;
 
-// Input UTXO is a tuple of txid, vout, amount, script
-typedef std::tuple<COutPoint, CAmount, CScript> MergeToAddressInputUTXO;
-
-// Input JSOP is a tuple of JSOutpoint, note, amount, spending key
-typedef std::tuple<JSOutPoint, SproutNote, CAmount, SproutSpendingKey> MergeToAddressInputSproutNote;
-
-typedef std::tuple<SaplingOutPoint, SaplingNote, CAmount, SaplingExpandedSpendingKey> MergeToAddressInputSaplingNote;
-
 // A recipient is a tuple of address, memo (optional if zaddr)
-typedef std::pair<libzcash::PaymentAddress, std::string> MergeToAddressRecipient;
+typedef std::pair<libzcash::PaymentAddress, std::optional<Memo>> MergeToAddressRecipient;
 
 // Package of info which is passed to perform_joinsplit methods.
 struct MergeToAddressJSInfo {
@@ -56,12 +50,11 @@ class AsyncRPCOperation_mergetoaddress : public AsyncRPCOperation
 {
 public:
     AsyncRPCOperation_mergetoaddress(
-        std::optional<TransactionBuilder> builder,
-        CMutableTransaction contextualTx,
-        std::vector<MergeToAddressInputUTXO> utxoInputs,
-        std::vector<MergeToAddressInputSproutNote> sproutNoteInputs,
-        std::vector<MergeToAddressInputSaplingNote> saplingNoteInputs,
+        WalletTxBuilder builder,
+        ZTXOSelector ztxoSelector,
+        SpendableInputs allInputs,
         MergeToAddressRecipient recipient,
+        TransactionStrategy strategy,
         CAmount fee = DEFAULT_FEE,
         UniValue contextInfo = NullUniValue);
     virtual ~AsyncRPCOperation_mergetoaddress();
@@ -71,6 +64,8 @@ public:
     AsyncRPCOperation_mergetoaddress(AsyncRPCOperation_mergetoaddress&&) = delete;                 // Move construct
     AsyncRPCOperation_mergetoaddress& operator=(AsyncRPCOperation_mergetoaddress const&) = delete; // Copy assign
     AsyncRPCOperation_mergetoaddress& operator=(AsyncRPCOperation_mergetoaddress&&) = delete;      // Move assign
+
+    tl::expected<void, InputSelectionError> prepare(const CChainParams& chainparams, CWallet& wallet);
 
     virtual void main();
 
@@ -85,15 +80,13 @@ private:
 
     UniValue contextinfo_; // optional data to include in return value from getStatus()
 
-    bool isUsingBuilder_; // Indicates that no Sprout addresses are involved
     uint32_t consensusBranchId_;
     CAmount fee_;
     int mindepth_;
     bool isToTaddr_;
     bool isToZaddr_;
-    CTxDestination toTaddr_;
     PaymentAddress toPaymentAddress_;
-    std::string memo_;
+    std::optional<Memo> memo_;
 
     ed25519::VerificationKey joinSplitPubKey_;
     ed25519::SigningKey joinSplitPrivKey_;
@@ -101,35 +94,14 @@ private:
     // The key is the result string from calling JSOutPoint::ToString()
     std::unordered_map<std::string, MergeToAddressWitnessAnchorData> jsopWitnessAnchorMap;
 
-    std::vector<MergeToAddressInputUTXO> utxoInputs_;
-    std::vector<MergeToAddressInputSproutNote> sproutNoteInputs_;
-    std::vector<MergeToAddressInputSaplingNote> saplingNoteInputs_;
+    WalletTxBuilder builder_;
+    ZTXOSelector ztxoSelector_;
+    SpendableInputs allInputs_;
+    TransactionStrategy strategy_;
 
-    TransactionBuilder builder_;
-    CTransaction tx_;
+    std::optional<TransactionEffects> effects_;
 
-    std::array<unsigned char, ZC_MEMO_SIZE> get_memo_from_hex_string(std::string s);
-    bool main_impl();
-
-    // JoinSplit without any input notes to spend
-    UniValue perform_joinsplit(MergeToAddressJSInfo&);
-
-    // JoinSplit with input notes to spend (JSOutPoints))
-    UniValue perform_joinsplit(MergeToAddressJSInfo&, std::vector<JSOutPoint>&);
-
-    // JoinSplit where you have the witnesses and anchor
-    UniValue perform_joinsplit(
-        MergeToAddressJSInfo& info,
-        std::vector<std::optional<SproutWitness>> witnesses,
-        uint256 anchor);
-
-    void lock_utxos();
-
-    void unlock_utxos();
-
-    void lock_notes();
-
-    void unlock_notes();
+    uint256 main_impl(CWallet& wallet, const CChainParams& chainparams);
 
     // payment disclosure!
     std::vector<PaymentDisclosureKeyInfo> paymentDisclosureData_;
@@ -144,44 +116,11 @@ public:
 
     TEST_FRIEND_AsyncRPCOperation_mergetoaddress(std::shared_ptr<AsyncRPCOperation_mergetoaddress> ptr) : delegate(ptr) {}
 
-    CTransaction getTx()
-    {
-        return delegate->tx_;
-    }
-
-    void setTx(CTransaction tx)
-    {
-        delegate->tx_ = tx;
-    }
-
     // Delegated methods
 
-    std::array<unsigned char, ZC_MEMO_SIZE> get_memo_from_hex_string(std::string s)
+    uint256 main_impl(CWallet& wallet, const CChainParams& chainparams)
     {
-        return delegate->get_memo_from_hex_string(s);
-    }
-
-    bool main_impl()
-    {
-        return delegate->main_impl();
-    }
-
-    UniValue perform_joinsplit(MergeToAddressJSInfo& info)
-    {
-        return delegate->perform_joinsplit(info);
-    }
-
-    UniValue perform_joinsplit(MergeToAddressJSInfo& info, std::vector<JSOutPoint>& v)
-    {
-        return delegate->perform_joinsplit(info, v);
-    }
-
-    UniValue perform_joinsplit(
-        MergeToAddressJSInfo& info,
-        std::vector<std::optional<SproutWitness>> witnesses,
-        uint256 anchor)
-    {
-        return delegate->perform_joinsplit(info, witnesses, anchor);
+        return delegate->main_impl(wallet, chainparams);
     }
 
     void set_state(OperationStatus state)