+++ src/wallet/gtest/test_rpc_wallet.cpp
@@ -17,8 +17,19 @@
 #include <librustzcash.h>
 #include <rust/ed25519.h>
 
+namespace {
+
 bool find_error(const UniValue& objError, const std::string& expected) {
     return find_value(objError, "message").get_str().find(expected) != string::npos;
+}
+
+CWalletTx FakeWalletTx() {
+    CMutableTransaction mtx;
+    mtx.vout.resize(1);
+    mtx.vout[0].nValue = 1;
+    return CWalletTx(nullptr, mtx);
+}
+
 }
 
 // TODO: test private methods
@@ -46,128 +57,26 @@ TEST(WalletRPCTests, RPCZMergeToAddressInternals)
     auto taddr = pwalletMain->GenerateNewKey(true).GetID();
     std::string taddr_string = keyIO.EncodeDestination(taddr);
 
-    MergeToAddressRecipient taddr1(keyIO.DecodePaymentAddress(taddr_string).value(), "");
+    MergeToAddressRecipient taddr1(keyIO.DecodePaymentAddress(taddr_string).value(), Memo());
     auto pa = pwalletMain->GenerateNewSproutZKey();
-    MergeToAddressRecipient zaddr1(pa, "DEADBEEF");
+    MergeToAddressRecipient zaddr1(pa, Memo());
+
+    WalletTxBuilder builder(Params(), minRelayTxFee);
+    auto selector = CWallet::LegacyTransparentZTXOSelector(
+            true,
+            TransparentCoinbasePolicy::Disallow);
+    TransactionStrategy strategy(PrivacyPolicy::AllowRevealedSenders);
 
     // Insufficient funds
     {
-        std::vector<MergeToAddressInputUTXO> inputs = { MergeToAddressInputUTXO{COutPoint{uint256(),0},0, CScript()} };
-        std::shared_ptr<AsyncRPCOperation> operation(new AsyncRPCOperation_mergetoaddress(std::nullopt, mtx, inputs, {}, {}, zaddr1) );
-        operation->main();
-        EXPECT_TRUE(operation->isFailed());
-        std::string msg = operation->getErrorMessage();
-        EXPECT_TRUE(msg.find("Insufficient funds, have 0.00 and miners fee is 0.00001") != string::npos);
-    }
-
-    // get_memo_from_hex_string())
-    {
-        std::vector<MergeToAddressInputUTXO> inputs = { MergeToAddressInputUTXO{COutPoint{uint256(),0},100000, CScript()} };
-        std::shared_ptr<AsyncRPCOperation> operation(new AsyncRPCOperation_mergetoaddress(std::nullopt, mtx, inputs, {}, {}, zaddr1) );
-        std::shared_ptr<AsyncRPCOperation_mergetoaddress> ptr = std::dynamic_pointer_cast<AsyncRPCOperation_mergetoaddress> (operation);
-        TEST_FRIEND_AsyncRPCOperation_mergetoaddress proxy(ptr);
-
-        std::string memo = "DEADBEEF";
-        std::array<unsigned char, ZC_MEMO_SIZE> array = proxy.get_memo_from_hex_string(memo);
-        EXPECT_EQ(array[0], 0xDE);
-        EXPECT_EQ(array[1], 0xAD);
-        EXPECT_EQ(array[2], 0xBE);
-        EXPECT_EQ(array[3], 0xEF);
-        for (int i=4; i<ZC_MEMO_SIZE; i++) {
-            EXPECT_EQ(array[i], 0x00);  // zero padding
-        }
-
-        // memo is longer than allowed
-        std::vector<char> v (2 * (ZC_MEMO_SIZE+1));
-        std::fill(v.begin(), v.end(), 'A');
-        std::string bigmemo(v.begin(), v.end());
-
-        try {
-            proxy.get_memo_from_hex_string(bigmemo);
-            FAIL() << "Should have caused an error";
-        } catch (const UniValue& objError) {
-            EXPECT_TRUE(find_error(objError, "too big"));
-        }
-
-        // invalid hexadecimal string
-        std::fill(v.begin(), v.end(), '@'); // not a hex character
-        std::string badmemo(v.begin(), v.end());
-
-        try {
-            proxy.get_memo_from_hex_string(badmemo);
-            FAIL() << "Should have caused an error";
-        } catch (const UniValue& objError) {
-            EXPECT_TRUE(find_error(objError, "hexadecimal format"));
-        }
-
-        // odd length hexadecimal string
-        std::fill(v.begin(), v.end(), 'A');
-        v.resize(v.size() - 1);
-        assert(v.size() % 2 == 1); // odd length
-        std::string oddmemo(v.begin(), v.end());
-        try {
-            proxy.get_memo_from_hex_string(oddmemo);
-            FAIL() << "Should have caused an error";
-        } catch (const UniValue& objError) {
-            EXPECT_TRUE(find_error(objError, "hexadecimal format"));
-        }
-    }
-
-    // Test the perform_joinsplit methods.
-    {
-        // Dummy input so the operation object can be instantiated.
-        std::vector<MergeToAddressInputUTXO> inputs = { MergeToAddressInputUTXO{COutPoint{uint256(),0},100000, CScript()} };
-        std::shared_ptr<AsyncRPCOperation> operation(new AsyncRPCOperation_mergetoaddress(std::nullopt, mtx, inputs, {}, {}, zaddr1) );
-        std::shared_ptr<AsyncRPCOperation_mergetoaddress> ptr = std::dynamic_pointer_cast<AsyncRPCOperation_mergetoaddress> (operation);
-        TEST_FRIEND_AsyncRPCOperation_mergetoaddress proxy(ptr);
-
-        // Enable test mode so tx is not sent and proofs are not generated
-        static_cast<AsyncRPCOperation_mergetoaddress *>(operation.get())->testmode = true;
-
-        MergeToAddressJSInfo info;
-        std::vector<std::optional < SproutWitness>> witnesses;
-        uint256 anchor;
-        try {
-            proxy.perform_joinsplit(info, witnesses, anchor);
-            FAIL() << "Should have caused an error";
-        } catch (const std::runtime_error & e) {
-            EXPECT_TRUE(string(e.what()).find("anchor is null") != string::npos);
-        }
-
-        try {
-            std::vector<JSOutPoint> v;
-            proxy.perform_joinsplit(info, v);
-            FAIL() << "Should have caused an error";
-        } catch (const std::runtime_error & e) {
-            EXPECT_TRUE(string(e.what()).find("anchor is null") != string::npos);
-        }
-
-        info.notes.push_back(SproutNote());
-        try {
-            proxy.perform_joinsplit(info);
-            FAIL() << "Should have caused an error";
-        } catch (const std::runtime_error & e) {
-            EXPECT_TRUE(string(e.what()).find("number of notes") != string::npos);
-        }
-
-        info.notes.clear();
-        info.vjsin.push_back(JSInput());
-        info.vjsin.push_back(JSInput());
-        info.vjsin.push_back(JSInput());
-        try {
-            proxy.perform_joinsplit(info);
-            FAIL() << "Should have caused an error";
-        } catch (const std::runtime_error & e) {
-            EXPECT_TRUE(string(e.what()).find("unsupported joinsplit input") != string::npos);
-        }
-
-        info.vjsin.clear();
-        try {
-            proxy.perform_joinsplit(info);
-            FAIL() << "Should have caused an error";
-        } catch (const std::runtime_error & e) {
-            EXPECT_TRUE(string(e.what()).find("error verifying joinsplit") != string::npos);
-        }
+        SpendableInputs inputs;
+        auto wtx = FakeWalletTx();
+        inputs.utxos.emplace_back(COutput(&wtx, 0, 100, true));
+        auto operation = AsyncRPCOperation_mergetoaddress(std::move(builder), selector, inputs, zaddr1, strategy, 0);
+        operation.main();
+        EXPECT_TRUE(operation.isFailed());
+        std::string msg = operation.getErrorMessage();
+        EXPECT_EQ(msg, "Sending funds into the Sprout pool is no longer supported.");
     }
     }
     UnloadGlobalWallet();